+++ src/data/bangkokStreetsData.js
@@ -1,0 +1,2 @@
+    export const bangkokStreetsData = [ { value: 'Sukhumvit Road', label: 'Sukhumvit Road', label_th: 'ถนนสุขุมวิท', alleys: [ { value: 'Sukhumvit Soi 1', label: 'Sukhumvit Soi 1', label_th: 'สุขุมวิท ซอย 1' }, { value: 'Sukhumvit Soi 3 (Soi Nana Nuea)', label: 'Sukhumvit Soi 3 (Nana Nuea)', label_th: 'สุขุมวิท ซอย 3 (นานาเหนือ)' }, { value: 'Sukhumvit Soi 11', label: 'Sukhumvit Soi 11', label_th: 'สุขุมวิท ซอย 11' }, { value: 'Sukhumvit Soi 21 (Soi Asok)', label: 'Sukhumvit Soi 21 (Asok)', label_th: 'สุขุมวิท ซอย 21 (อโศก)' }, ], associatedKhwaengs: ['Khlong Toei Nuea', 'Phrom Phong', 'Thong Lo (Watthana)', 'Ekkamai (Watthana)', 'Khlong Tan Nuea', 'Khlong Toei (Khwaeng)', 'Phra Khanong (Khlong Toei)'] }, { value: 'Lat Phrao Road', label: 'Lat Phrao Road', label_th: 'ถนนลาดพร้าว', alleys: [ { value: 'Lat Phrao Soi 1', label: 'Lat Phrao Soi 1', label_th: 'ลาดพร้าว ซอย 1' }, { value: 'Lat Phrao Soi 71', label: 'Lat Phrao Soi 71', label_th: 'ลาดพร้าว ซอย 71' }, { value: 'Lat Phrao Soi 101', label: 'Lat Phrao Soi 101', label_th: 'ลาดพร้าว ซอย 101' }, ], associatedKhwaengs: ['Lat Phrao (Khwaeng)', 'Chorakhe Bua'] }, { value: 'Silom Road', label: 'Silom Road', label_th: 'ถนนสีลม', alleys: [ { value: 'Silom Soi 1', label: 'Silom Soi 1', label_th: 'สีลม ซอย 1' }, { value: 'Silom Soi 9', label: 'Silom Soi 9', label_th: 'สีลม ซอย 9' }, ], associatedKhwaengs: ['Silom', 'Suriyawong', 'Si Phraya'] }, ];
+
