+++ src/data/categories.js
@@ -1,0 +1,2 @@
+    export const categories = [ { value: 'all', label: 'All Categories', label_th: 'ทุกหมวดหมู่', icon: '🌐', isTopLevel: true }, { value: 'Online', label: 'Online', label_th: 'ออนไลน์', icon: '💻', isTopLevel: true, platforms: [ { value: 'Facebook', label: 'Facebook', label_th: 'เฟซบุ๊ก', subcategories: [ { value: 'Influencer', label: 'Influencer', label_th: 'อินฟลูเอนเซอร์' }, { value: 'Shop', label: 'Shop', label_th: 'ร้านค้า' }, { value: 'Page', label: 'Page', label_th: 'เพจ' }, { value: 'Group', label: 'Group', label_th: 'กลุ่ม' }, ] }, { value: 'Shopee', label: 'Shopee', label_th: 'ช้อปปี้', subcategories: [ { value: 'Shop', label: 'Shop', label_th: 'ร้านค้า' }, { value: 'Seller', label: 'Seller', label_th: 'ผู้ขาย' }, ] }, { value: 'Instagram', label: 'Instagram', label_th: 'อินสตาแกรม', subcategories: [ { value: 'Influencer', label: 'Influencer', label_th: 'อินฟลูเอนเซอร์' }, { value: 'Shop', label: 'Shop', label_th: 'ร้านค้า' }, ] }, { value: 'TikTok', label: 'TikTok', label_th: 'ติ๊กต็อก', subcategories: [ { value: 'Influencer', label: 'Influencer', label_th: 'อินฟลูเอนเซอร์' }, { value: 'Shop', label: 'Shop', label_th: 'ร้านค้า' }, ] }, { value: 'YouTube', label: 'YouTube', label_th: 'ยูทูป', subcategories: [ { value: 'Channel', label: 'Channel', label_th: 'ช่อง' }, { value: 'Content Creator', label: 'Content Creator', label_th: 'ผู้สร้างเนื้อหา' }, ] }, { value: 'Telegram', label: 'Telegram', label_th: 'เทเลแกรม', subcategories: [ { value: 'Group', label: 'Group', label_th: 'กลุ่ม' }, { value: 'Channel', label: 'Channel', label_th: 'ช่อง' }, ] }, { value: 'Line', label: 'Line', label_th: 'ไลน์', subcategories: [ { value: 'Official Account', label: 'Official Account', label_th: 'บัญชีทางการ' }, { value: 'Group', label: 'Group', label_th: 'กลุ่ม' }, ] }, { value: 'Website', label: 'Website', label_th: 'เว็บไซต์', subcategories: [ { value: 'E-commerce', label: 'E-commerce', label_th: 'อีคอมเมิร์ซ' }, { value: 'Blog', label: 'Blog', label_th: 'บล็อก' }, { value: 'Forum', label: 'Forum', label_th: 'ฟอรั่ม' }, { value: 'Service Provider', label: 'Service Provider', label_th: 'ผู้ให้บริการ' }, ] }, { value: 'Online Service (General)', label: 'Online Service (General)', label_th: 'บริการออนไลน์ (ทั่วไป)' }, ] }, { value: 'Real World', label: 'Real World', label_th: 'โลกจริง', icon: '🌍', isTopLevel: true, subcategories: [ { value: 'Local Services', label: 'Local Services', label_th: 'บริการท้องถิ่น', icon: '🛠️', subcategories: [ { value: 'Movers', label: 'Movers', label_th: 'ผู้ให้บริการขนย้าย' }, { value: 'Tutors', label: 'Tutors', label_th: 'ติวเตอร์' }, { value: 'Handymen', label: 'Handymen', label_th: 'ช่างซ่อมบำรุง' }, { value: 'Cleaners', label: 'Cleaners', label_th: 'ผู้ให้บริการทำความสะอาด' }, { value: 'Plumbers', label: 'Plumbers', label_th: 'ช่างประปา' }, { value: 'Electricians', label: 'Electricians', label_th: 'ช่างไฟฟ้า' }, { value: 'Mechanics', label: 'Mechanics', label_th: 'ช่างยนต์' }, ] }, { value: 'Facebook Marketplace', label: 'Facebook Marketplace', label_th: 'ตลาด Facebook', icon: '🛒', subcategories: [ { value: 'Used Cars', label: 'Used Cars', label_th: 'รถยนต์มือสอง' }, { value: 'Used Furniture', label: 'Used Furniture', label_th: 'เฟอร์นิเจอร์มือสอง' }, { value: 'Used Electronics', label: 'Used Electronics', label_th: 'อุปกรณ์อิเล็กทรอนิกส์มือสอง' }, ] }, { value: 'Goods & Products', label: 'Goods & Products', label_th: 'สินค้าและผลิตภัณฑ์', icon: '📦', subcategories: [ { value: 'Used Clothing', label: 'Used Clothing', label_th: 'เสื้อผ้ามือสอง' }, { value: 'Used Electronics', label: 'Used Electronics', label_th: 'อุปกรณ์อิเล็กทรอนิกส์มือสอง' }, { value: 'Furniture', label: 'Furniture', label_th: 'เฟอร์นิเจอร์' }, { value: 'Antiques', label: 'Antiques', label_th: 'ของเก่า' }, { value: 'Collectibles', label: 'Collectibles', label_th: 'ของสะสม' }, ] }, { value: 'Landlords', label: 'Landlords', label_th: 'เจ้าของที่ดิน/อาคาร', icon: '🏠', subcategories: [ { value: 'Apartment', label: 'Apartment', label_th: 'อพาร์ตเมนต์' }, { value: 'Condo', label: 'Condo', label_th: 'คอนโด' }, { value: 'House', label: 'House', label_th: 'บ้าน' }, ] }, { value: 'Restaurants & Cafes', label: 'Restaurants & Cafes', label_th: 'ร้านอาหารและคาเฟ่', icon: '🍽️', subcategories: [ { value: 'Thai Food', label: 'Thai Food', label_th: 'อาหารไทย' }, { value: 'Japanese Food', label: 'Japanese Food', label_th: 'อาหารญี่ปุ่น' }, { value: 'Italian Food', label: 'Italian Food', label_th: 'อาหารอิตาเลียน' }, { value: 'Cafe', label: 'Cafe', label_th: 'คาเฟ่' }, { value: 'Street Food Stall', label: 'Street Food Stall', label_th: 'ร้านอาหารริมทาง' }, { value: 'Fine Dining', label: 'Fine Dining', label_th: 'ร้านอาหารหรู' }, ] }, { value: 'Retail Stores', label: 'Retail Stores', label_th: 'ร้านค้าปลีก', icon: '🛍️', subcategories: [ { value: 'Fashion Boutique', label: 'Fashion Boutique', label_th: 'ร้านบูติกเสื้อผ้า' }, { value: 'Electronics Store', label: 'Electronics Store', label_th: 'ร้านเครื่องใช้ไฟฟ้า' }, { value: 'Supermarket', label: 'Supermarket', label_th: 'ซูเปอร์มาร์เก็ต' }, { value: 'Convenience Store', label: 'Convenience Store', label_th: 'ร้านสะดวกซื้อ' }, ] }, { value: 'Education Centers', label: 'Education Centers', label_th: 'ศูนย์การศึกษา', icon: '🎓', subcategories: [ { value: 'Language School', label: 'Language School', label_th: 'โรงเรียนสอนภาษา' }, { value: 'Tutoring Center', label: 'Tutoring Center', label_th: 'ศูนย์กวดวิชา' }, { value: 'Workshop/Classes', label: 'Workshop/Classes', label_th: 'เวิร์คช็อป/คลาสเรียน' }, ] }, { value: 'Healthcare Services', label: 'Healthcare Services', label_th: 'บริการด้านสุขภาพ', icon: '🏥', subcategories: [ { value: 'Clinic', label: 'Clinic', label_th: 'คลินิก' }, { value: 'Hospital', label: 'Hospital', label_th: 'โรงพยาบาล' }, { value: 'Dentist', label: 'Dentist', label_th: 'ทันตแพทย์' }, { value: 'Pharmacy', label: 'Pharmacy', label_th: 'ร้านขายยา' }, ] }, ] } ];
+
