+++ src/data/citiesData.js
@@ -1,0 +1,2 @@
+    export const citiesData = [ { value: 'Bangkok', label: 'Bangkok', label_th: 'กรุงเทพมหานคร', districts: [ { value: 'Sathorn', label: 'Sathorn', label_th: 'สาทร', zones: [ { value: 'Sathorn-Silom CBD', label: 'Sathorn-Silom CBD', label_th: 'สาทร-สีลม ศูนย์กลางธุรกิจ', khwaengValues: ['Thung Wat Don', 'Thung Maha Mek', 'Yan Nawa (Sathorn)', 'Chong Nonsi (Sathorn)', 'Suan Phlu'] }, { value: 'Riverside Area (Sathorn)', label: 'Riverside Area (Sathorn)', label_th: 'พื้นที่ริมแม่น้ำ (สาทร)', khwaengValues: ['Thung Wat Don', 'Yan Nawa (Sathorn)'] }, { value: 'Naradhiwas Area (Sathorn)', label: 'Naradhiwas Area (Sathorn)', label_th: 'พื้นที่นราธิวาส (สาทร)', khwaengValues: ['Thung Wat Don', 'Chong Nonsi (Sathorn)', 'Thung Maha Mek'] }, ], subDistricts: [ { value: 'Thung Wat Don', label: 'Thung Wat Don', label_th: 'ทุ่งวัดดอน' }, { value: 'Yan Nawa (Sathorn)', label: 'Yan Nawa', label_th: 'ยานนาวา' }, { value: 'Thung Maha Mek', label: 'Thung Maha Mek', label_th: 'ทุ่งมหาเมฆ' }, { value: 'Chong Nonsi (Sathorn)', label: 'Chong Nonsi', label_th: 'ช่องนนทรี' }, { value: 'Suan Phlu', label: 'Suan Phlu', label_th: 'สวนพลู' }, ] }, { value: 'Watthana', label: 'Watthana', label_th: 'วัฒนา', zones: [ { value: 'Sukhumvit Main Area (Watthana)', label: 'Sukhumvit Main Area', label_th: 'สุขุมวิทพื้นที่หลัก (วัฒนา)', khwaengValues: ['Khlong Toei Nuea', 'Phrom Phong', 'Thong Lo (Watthana)', 'Ekkamai (Watthana)', 'Khlong Tan Nuea'] }, { value: 'Asok-Nana Area', label: 'Asok-Nana Area', label_th: 'อโศก-นานา', khwaengValues: ['Khlong Toei Nuea'] }, { value: 'Phrom Phong - Thong Lo', label: 'Phrom Phong - ทองหล่อ', label_th: 'พร้อมพงษ์ - ทองหล่อ', khwaengValues: ['Phrom Phong', 'Thong Lo (Watthana)'] }, { value: 'Ekkamai Area', label: 'Ekkamai Area', label_th: 'เอกมัย', khwaengValues: ['Ekkamai (Watthana)', 'Khlong Tan Nuea'] }, ], subDistricts: [ { value: 'Khlong Toei Nuea', label: 'Khlong Toei Nuea', label_th: 'คลองเตยเหนือ' }, { value: 'Phrom Phong', label: 'Phrom Phong', label_th: 'พร้อมพงษ์' }, { value: 'Thong Lo (Watthana)', label: 'Thong Lo', label_th: 'ทองหล่อ' }, { value: 'Ekkamai (Watthana)', label: 'Ekkamai', label_th: 'เอกมัย' }, { value: 'Khlong Tan Nuea', label: 'Khlong Tan Nuea', label_th: 'คลองตันเหนือ' }, ] }, { value: 'Khlong Toei', label: 'Khlong Toei', label_th: 'คลองเตย', zones: [ { value: 'Sukhumvit Extension (Khlong Toei)', label: 'Sukhumvit Extension', label_th: 'สุขุมวิทส่วนขยาย (คลองเตย)', khwaengValues: ['Khlong Toei (Khwaeng)', 'Khlong Tan (Khlong Toei)', 'Phra Khanong (Khlong Toei)', 'Lumphini (Khlong Toei)'] }, { value: 'Port Area (Khlong Toei)', label: 'Port Area', label_th: 'พื้นที่ท่าเรือ (คลong Toei)', khwaengValues: ['Khlong Toei (Khwaeng)'] }, { value: 'Convention Center Area', label: 'Convention Center Area', label_th: 'พื้นที่ศูนย์ประชุม', khwaengValues: ['Khlong Toei (Khwaeng)'] }, ], subDistricts: [ { value: 'Khlong Toei (Khwaeng)', label: 'Khlong Toei', label_th: 'คลองเตย' }, { value: 'Khlong Tan (Khlong Toei)', label: 'Khlong Tan', label_th: 'คลองตัน' }, { value: 'Phra Khanong (Khlong Toei)', label: 'Phra Khanong', label_th: 'พระโขนง' }, { value: 'Lumphini (Khlong Toei)', label: 'Lumphini', label_th: 'ลุมพินี' }, ] }, { value: 'Lat Phrao', label: 'Lat Phrao', label_th: 'ลาดพร้าว', zones: [ { value: 'Ladprao Main Road Area', label: 'Ladprao Main Road Area', label_th: 'พื้นที่ถนนลาดพร้าวสายหลัก', khwaengValues: ['Lat Phrao (Khwaeng)'] }, { value: 'Ladprao Soi Areas', label: 'Ladprao Soi Areas', label_th: 'พื้นที่ซอยลาดพร้าว', khwaengValues: ['Lat Phrao (Khwaeng)', 'Chorakhe Bua'] }, { value: 'Chok Chai 4 Area', label: 'Chok Chai 4 Area', label_th: 'พื้นที่โชคชัย 4', khwaengValues: ['Lat Phrao (Khwaeng)'] }, ], subDistricts: [ { value: 'Lat Phrao (Khwaeng)', label: 'Lat Phrao', label_th: 'ลาดพร้าว' }, { value: 'Chorakhe Bua', label: 'Chorakhe Bua', label_th: 'จระเข้บัว' }, ] }, ] }, { value: 'Online', label: 'Online', label_th: 'ออนไลน์', districts: [] }, ];
+
