+++ src/utils/auth.js
@@ -1,0 +1,4 @@
+import { signInWithGoogle, signOut as firebaseSignOut } from './mockFirebase';
+
+export const signIn = signInWithGoogle;
+export const signOut = firebaseSignOut;
